[lldb] Move RegisterInfo and RegisterSet into their own header
RegisterInfo holds onto a pointer to RegisterFlags which is in
lldbUtility. My understanding of lldb-private-types is that it contains
types used throughout LLDB but should generally stand on their own (i.e.
without dependencies on specific lldb types and headers). Thus I am
moving it to lldbUtility.

RegisterSet _does_ stand on its own but it is pretty strongly coupled to
RegisterInfo.
